Adding func test project for NuGet.Commands
