+++ test/NuGet.Core.FuncTests/NuGet.Commands.FuncTest/project.json
@@ -1,0 +1,18 @@
+﻿{
+  "version": "3.4.0-*",
+  "dependencies": {
+    "NuGet.Commands": "3.4.0-*",
+    "NuGet.Test.Utility": "3.4.0-*",
+    "xunit": "2.1.0",
+    "xunit.runner.dnx": "2.1.0-rc1-build204"
+  },
+  "commands": {
+    "test": "xunit.runner.dnx"
+  },
+  "frameworks": {
+    "dnx451": { },
+    "dnxcore50": {
+      "imports": [ "portable-net45+win8" ]
+    }
+  }
+}
