+++ test/NuGet.Core.FuncTests/NuGet.Commands.FuncTest/xunit.runner.json
@@ -1,0 +1,4 @@
+﻿{
+  "maxParallelThreads": 1,
+  "parallelizeTestCollections": false
+}
