+++ test/NuGet.Core.FuncTests/NuGet.Config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <packageSources>
+  </packageSources>
+  <disabledPackageSources>
+     <add key="BuildFeed" value="Nupkgs" />
+  </disabledPackageSources>
+</configuration>
