+++ test/NuGet.Core.FuncTests/global.json
@@ -1,0 +1,6 @@
+{
+  "projects": [
+    "../../src/NuGet.Core",
+    "../../lib"
+  ]
+}
